qml: conditionally load the banner in the OnboardingInfo control
refactors the banner property in OnboardingInfo from an alias to a bool
which signals when to have the banner_loader as active. Introduces
the bannerItem as the new property which holds the banner as an alias;
updates banner definitions to set bannerItem.

Github-Pull: bitcoin-core/gui-qml#183
Rebased-From: 25be80c433d218f3293dbc48295378c12b4dd80c