qml: introduce ChainModel
The ChainModel is responsible for managing the block information
that the gui components need to show. In this case, the ChainModel
will provide a list of ratios for the block times of the last 12 hours
to be rendered on the block clock's dial.

Github-Pull: bitcoin-core/gui-qml#220
Rebased-From: 60bd8cb73acdd650b3b9f03693896cfd4573cc9c

Co-authored-by: shaavan <shaavan.github@gmail.com>